+++ pg_documentdb/include/utils/feature_counter.h
@@ -19,7 +19,7 @@
 #include <port/atomics.h>
 
 #define MAX_FEATURE_NAME_LENGTH 255
-#define MAX_FEATURE_COUNT 291
+#define MAX_FEATURE_COUNT 292
 
 /* Internal features that are not exposed */
 #define INTERNAL_FEATURE_TYPE MAX_FEATURE_COUNT
@@ -276,6 +276,9 @@ typedef enum
 	FEATURE_QUERY_OPERATOR_NEARSPHERE,
 	FEATURE_QUERY_OPERATOR_SAMPLERATE,
 	FEATURE_QUERY_OPERATOR_TEXT,
+
+	/* Feature mapping region - Role CRUD */
+	FEATURE_ROLE_CREATE,
 
 	/* Feature counter region - Aggregation stages */
 	FEATURE_STAGE_ADD_FIELDS,