+++ pg_documentdb/sql/documentdb--0.105-0--0.106-0.sql
@@ -9,5 +9,9 @@
 #include "udfs/metadata/collection--0.106-0.sql"
 #include "udfs/metadata/empty_data_table--0.106-0.sql"
 
+#include "udfs/roles/create_role--0.106-0.sql"
+#include "udfs/roles/drop_role--0.106-0.sql"
+#include "udfs/roles/roles_info--0.106-0.sql"
+#include "udfs/roles/update_role--0.106-0.sql"
 #include "udfs/aggregation/group_aggregates_support--0.106-0.sql"
 #include "udfs/aggregation/group_aggregates--0.106-0.sql"